requirements.txt
# requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
google-generativeai>=0.6.0
python-multipart>=0.0.6
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.3.0
cryptography>=41.0.3